Main and Pause Menu Finished
Main menu has animation and intro music.
Pause menu has back to main menu and exit game options.
Changed the camera view in the level so it doesn't look strange after it's built.
Deleted a couple of tiles leading to the bedroom.

(Menus may look strange in the Unity project, but they look fine in the build)